feat: 게시물 수정, 삭제 예외처리(#6)

+++ backend/src/main/java/com/miniprac/board/service/BoardService.java
@@ -6,19 +6,15 @@ import com.miniprac.board.domain.BoardCategoryType;
 import com.miniprac.board.domain.repository.BoardCategoryRepository;
 import com.miniprac.board.domain.repository.BoardRepository;
 import com.miniprac.board.dto.BoardRequest;
-import com.miniprac.board.dto.BoardResponse;
 import com.miniprac.board.service.exception.BoardCategoryNotFoundException;
 import com.miniprac.board.service.exception.BoardNotFoundException;
 import com.miniprac.common.dto.PageDto;
+import com.miniprac.common.exception.PermissionException;
+import com.miniprac.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 @Transactional
@@ -45,15 +41,24 @@ public class BoardService {
         return boardRepository.findById(boardId).orElseThrow(BoardNotFoundException::new);
     }
 
-    public Board update(BoardRequest.Update request, Long boardId){
-        BoardCategory category = boardCategoryRepository.findByType(BoardCategoryType.from(request.getCategory())).orElseThrow(BoardCategoryNotFoundException::new);
+    public Board update(BoardRequest.Update request, Long boardId, User user){
+
         Board board = boardRepository.findById(boardId).orElseThrow(BoardNotFoundException::new);
+
+        if(!user.getId().equals(board.getCreatedBy().getId())){
+            throw new PermissionException();
+        }
+       BoardCategory category = boardCategoryRepository.findByType(BoardCategoryType.from(request.getCategory())).orElseThrow(BoardCategoryNotFoundException::new);
+
         board.update(request, category);
         return board;
     }
 
-    public void deleteBoard(Long boardId){
+    public void deleteBoard(Long boardId, User user){
         Board board = boardRepository.findById(boardId).orElseThrow(BoardNotFoundException::new);
+        if(!user.getId().equals(board.getCreatedBy().getId())){
+            throw new PermissionException();
+        }
         boardRepository.delete(board);
     }
